Open SVG files in image previewer by default
SVG images were opened in the XML text editor by default. This change
opens them in the image previewer instead. Assuming that most SVG images
will be created and exported from graphics software like Adobe
Illustrator or Inkscape, previewing the file content is the better
default.

Also changes the behaviour of `editors#register()` in that the latest
registered editor is used with the highest instead of lowest priority.