drops buggy and redundant afterRmdir handler
Deleting a directory causes afterRmdir's handler to be fired, and
therefore unloading tabs whose paths start with the directory's
path, mistakenly considering the files open in such tabs to be under
that directory.

There are cases where a tab's path starts with the directory's path,
and yet the file is not under the directory (e.g., /foo.c and /foo),
so this causes tabs to be closed and unloaded when they shouldn't.

This could be easily solved, except that there's a race condition
caused when the watcher.gui plugin asks the user, possibly too late,
whether to keep the tab of a deleted file open.

By removing this handler, the tab is closed and unloaded, only if the
user doesn't want to keep it open.